// src/__tests__/roll-a-dice/v0_0_1/RollADice.types.ts
/**
* This file was automatically generated by @cosmwasm/ts-codegen@1.12.1.
* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
*/

export type Addr = string;
export interface InstantiateMsg {
  config: ConfigMsg;
}
export interface ConfigMsg {
  accepted_denom: string;
  disabled: boolean;
  fee_percentage: string;
  gas_limit: string;
  max_bet: string;
  min_bet: string;
  operator?: Addr | null;
  random_cw_address: Addr;
}
export type ExecuteMsg = {
  set_config: ConfigMsg;
} | {
  play_request: PlayRequestMsg;
} | {
  receive_randomness: ReceiveRandomnessMsg;
} | {
  fetch_random_c_w_config: FetchRandomCWConfigMsg;
} | {
  withdraw: WithdrawMsg;
};
export type GameType = {
  exact_number: ExactNumberGame;
} | {
  high_low: HighLowGame;
};
export type Uint64 = string;
export type JobResult = {
  u8: number[];
} | {
  u16: number[];
} | {
  u32: number[];
} | {
  u64: Uint64[];
} | {
  u128: Uint128[];
} | {
  i8: number[];
} | {
  i16: number[];
} | {
  i32: number[];
} | {
  i64: Int64[];
} | {
  i128: Int128[];
};
export type Uint128 = string;
export type Int64 = string;
export type Int128 = string;
export type Token = {
  denom: string;
} | {
  address: Addr;
};
export interface PlayRequestMsg {
  game_type: GameType;
}
export interface ExactNumberGame {
  chosen_number: number;
}
export interface HighLowGame {
  dice_number: number;
}
export interface ReceiveRandomnessMsg {
  id: Uint64;
  results: JobResult[];
}
export interface FetchRandomCWConfigMsg {}
export interface WithdrawMsg {
  amount: string;
  token: Token;
}
export type QueryMsg = {
  config: {};
} | {
  query_game: {
    id: string;
  };
} | {
  user_games_query: UserGamesQueryMsg;
};
export interface UserGamesQueryMsg {
  cursor?: number | null;
  limit: number;
  user_address: Addr;
}
export type ConfigResponse = Config;
export interface Config {
  accepted_denom: string;
  disabled: boolean;
  fee_percentage: Uint64;
  gas_limit: Uint64;
  max_bet: Uint128;
  min_bet: Uint128;
  operator?: Addr | null;
  random_cw_address: Addr;
}
export type QueryPlayResponse = GameRequest;
export type GameStatus = "requested" | "won" | "lost" | "error" | "refunded";
export interface GameRequest {
  bet_amount: Uint128;
  bet_currency: string;
  created_at: Uint64;
  dice_results?: number[] | null;
  error_message?: string | null;
  game_type: GameType;
  id: Uint64;
  potential_winning_amount: Uint128;
  randomness_request_id?: string | null;
  randomness_serving_block_height?: string | null;
  status: GameStatus;
  updated_at: Uint64;
  user_address: Addr;
}
export interface QueryUserGamesResponse {
  games: GameRequest[];
  next_cursor?: Uint64 | null;
}